fix: (organização e desvio padrão)

+++ app/src/main/java/com/avaliadorimovel/details/interfaces/InterfaceDetailsView.kt
@@ -4,5 +4,5 @@ interface InterfaceDetailsView {
     fun setControls()
     fun createSamples()
     fun onValidationError()
-    fun navigateToResult(result: Float)
+    fun navigateToResult(result: Double)
 }
